Added Auto TPM Option
Got lazy and tired of opening the menu everytime I need to tpm while testing code. Do not know if this is something you are interested in adding, but since I made it figured I would open the PR to offer. It adds a auto tpm option to the teleport menu, which when enabled will automatically teleport you to a waypoint you set. Translations all done. Tested and works fine.